requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
python-multipart
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
bcrypt
email-validator
Pillow

# Dev/tests + IDE stubs: pip install -r requirements-dev.txt
